Handle Share action in What's New
