+++ iOS/DuckDuckGo/Actions/ShareAction+UIActivity.swift
@@ -1,0 +1,66 @@
+//
+//  ShareAction+UIActivity.swift
+//  DuckDuckGo
+//
+//  Copyright © 2025 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Foundation
+import SwiftUI
+import LinkPresentation
+
+class TitledURLActivityItem: NSObject, UIActivityItemSource {
+
+    let url: URL
+    let title: String
+
+    init(_ url: URL, _ title: String) {
+        self.url = url
+        self.title = title
+    }
+
+    func activityViewControllerPlaceholderItem(_ activityViewController: UIActivityViewController) -> Any {
+        url
+    }
+
+    func activityViewController(_ activityViewController: UIActivityViewController, itemForActivityType activityType: UIActivity.ActivityType?) -> Any? {
+        url
+    }
+
+    func activityViewControllerLinkMetadata(_ activityViewController: UIActivityViewController) -> LPLinkMetadata? {
+        let metadata = LPLinkMetadata()
+        metadata.title = title
+        metadata.url = url
+        return metadata
+    }
+
+}
+
+struct TitleValueShareItem: Identifiable {
+    var id: String {
+        value
+    }
+
+    var item: Any {
+        if let url = URL(string: value), let title = title {
+            return TitledURLActivityItem(url, title)
+        } else {
+            return value
+        }
+    }
+
+    let value: String
+    let title: String?
+}
